+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+
+## 0.5.0
+  * Update metadata generation to include parent-tap-stream-id for substreams
+  * Add forced-replication-method support [#27](https://github.com/singer-io/tap-ordway/pull/27)
+
+
 # 0.4.6
   * Bump requests to 2.33.0 for security updates [#28](https://github.com/singer-io/tap-ordway/pull/28)
 
@@ -9,6 +15,7 @@
   * Charges, Invoices, Orders, Payments, Plans, Refunds, Subscriptions, Billing runs, Billing Schedules, Chart of accounts, Credits, Customers, Payment runs, Products, Revenue Schedules, Webhooks
 
   * https://github.com/singer-io/tap-ordway/pull/21
+
 
 ## 0.4.4
   * Add replication-keys and primary-keys in the schema of journal_entries. [#26](https://github.com/singer-io/tap-ordway/pull/26)